+++ pkg/zfs/cmd/v1alpha1/zpool/add/predicate.go
@@ -43,6 +43,13 @@ func IsPoolSet() PredicateFunc {
 // IsDeviceTypeSet method check if the DeviceType field of PoolExpansion object is set.
 func IsDeviceTypeSet() PredicateFunc {
 	return func(p *PoolExpansion) bool {
+		// If no device type is mentioned in command
+		// ZFS consider it as a stripe type by itself
+		// We don't need to provide `stripe` as a type in command
+		// So If user has provided type `stripe`, we will ignore it
+		if p.DeviceType == "stripe" {
+			return false
+		}
 		return len(p.DeviceType) != 0
 	}
 }
@@ -50,6 +57,13 @@ func IsDeviceTypeSet() PredicateFunc {
 // IsTypeSet method check if the Type field of PoolExpansion object is set.
 func IsTypeSet() PredicateFunc {
 	return func(p *PoolExpansion) bool {
+		// If no device type is mentioned in command
+		// ZFS consider it as a stripe type by itself
+		// We don't need to provide `stripe` as a type in command
+		// So If user has provided type `stripe`, we will ignore it
+		if p.Type == "stripe" {
+			return false
+		}
 		return len(p.Type) != 0
 	}
 }
